bug fixes with classifier and its use
the script can now be called anywhere,
and the classifier on the same directory as
the script will be correctly used.

the probability for positive or negative messages
are now correctly extracted from the classifier,
rather than using a hacky method to compute.

negations in sentences are now handled
by inverting the positivity detected through
the classifier and lowering its magnitude.